requirements.txt
flask
flask-cors
numpy
pandas
faiss-gpu
langchain
langchain-community
langchain-core
sentence-transformers
rank-bm25
konlpy
ollama
transformers
torch